+++ src/steps/resource-manager/security/index.test.ts
@@ -2,6 +2,7 @@ import {
   fetchAssessments,
   fetchSecurityCenterContacts,
   fetchSecurityCenterPricingConfigurations,
+  fetchSecurityCenterSettings,
 } from '.';
 import { Recording } from '@jupiterone/integration-sdk-testing';
 import { IntegrationConfig } from '../../../types';
@@ -15,7 +16,7 @@ import { SecurityEntities, SecurityRelationships } from './constants';
 import { createMockAzureStepExecutionContext } from '../../../../test/createMockAzureStepExecutionContext';
 import { configFromEnv } from '../../../../test/integrationInstanceConfig';
 import { getMockAccountEntity } from '../../../../test/helpers/getMockEntity';
-import { v4 as uuid } from 'uuid';
+import { getConfigForTest } from './__tests__/utils';
 
 let recording: Recording;
 
@@ -128,27 +129,6 @@ describe('rm-security-center-pricing-configs', () => {
     return { accountEntity, subscriptionEntity };
   }
 
-  function getConfigForTest(config: IntegrationConfig): IntegrationConfig {
-    // The Azure Subscription Client has some validation before sending requests to
-    // this endpoint that requires "subscriptionId" to be a UUID.
-    //
-    // When running tests in CI, we set the value of `config.subscriptionId` to "subscriptionId",
-    // causing the Azure SDK to throw the following:
-    //
-    // "subscriptionId" with value "subscriptionId" should satisfy the constraint
-    // "Pattern": /^[0-9A-Fa-f]{8}-([0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$/.
-
-    const subscriptionIdValidationRegex = /^[0-9A-Fa-f]{8}-([0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$/;
-    return {
-      ...config,
-      subscriptionId: subscriptionIdValidationRegex.test(
-        config.subscriptionId || '',
-      )
-        ? config.subscriptionId
-        : uuid(),
-    };
-  }
-
   test('success', async () => {
     const config = getConfigForTest(configFromEnv);
 
@@ -190,6 +170,69 @@ describe('rm-security-center-pricing-configs', () => {
         properties: {
           _type: {
             const: SecurityRelationships.SUBSCRIPTION_HAS_PRICING_CONFIG._type,
+          },
+        },
+      },
+    });
+  });
+});
+
+describe('rm-security-center-settings', () => {
+  function getSetupEntities(config: IntegrationConfig) {
+    const accountEntity = getMockAccountEntity(config);
+
+    const subscriptionId = `/subscriptions/${config.subscriptionId}`;
+    const subscriptionEntity = {
+      _key: subscriptionId,
+      _type: 'subscription-type',
+      _class: 'subscription-class',
+    };
+
+    return { accountEntity, subscriptionEntity };
+  }
+
+  test('success', async () => {
+    const config = getConfigForTest(configFromEnv);
+
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'rm-security-center-settings',
+      options: {
+        matchRequestsBy: getMatchRequestsBy({ config }),
+      },
+    });
+
+    const { accountEntity, subscriptionEntity } = getSetupEntities(config);
+
+    const context = createMockAzureStepExecutionContext({
+      instanceConfig: config,
+      entities: [subscriptionEntity],
+      setData: {
+        [ACCOUNT_ENTITY_TYPE]: accountEntity,
+      },
+    });
+
+    await fetchSecurityCenterSettings(context);
+
+    const securityCenterSettingEntities = context.jobState.collectedEntities;
+
+    expect(securityCenterSettingEntities.length).toBeGreaterThan(0);
+    expect(securityCenterSettingEntities).toMatchGraphObjectSchema({
+      _class: SecurityEntities.SETTING._class,
+      schema: SecurityEntities.SETTING.schema,
+    });
+
+    const subscriptionSettingRelationships =
+      context.jobState.collectedRelationships;
+
+    expect(subscriptionSettingRelationships.length).toBe(
+      securityCenterSettingEntities.length,
+    );
+    expect(subscriptionSettingRelationships).toMatchDirectRelationshipSchema({
+      schema: {
+        properties: {
+          _type: {
+            const: SecurityRelationships.SUBSCRIPTION_HAS_SETTING._type,
           },
         },
       },